+++ include/UDRefl/Util.h
@@ -211,7 +211,7 @@ namespace Ubpa::UDRefl {
 	// - Func Obj::* : Func isn't && (const && is ok)
 	// - Func*
 	// [result]
-	// - type : ({const?} void* obj, void* result_buffer, ArgPtrBuffer argptr_buffer) -> Destructor
+	// - type : void({const?} void* obj, void* result_buffer, ArgPtrBuffer argptr_buffer)
 	// - size : 1
 	template<auto func_ptr>
 	constexpr auto wrap_member_function() noexcept;
@@ -219,7 +219,7 @@ namespace Ubpa::UDRefl {
 	// [func_ptr]
 	// - Func*
 	// [result]
-	// - type : (void* result_buffer, ArgPtrBuffer argptr_buffer) -> Destructor
+	// - type : void(void* result_buffer, ArgPtrBuffer argptr_buffer)
 	// - size : 1
 	template<auto func_ptr>
 	constexpr auto wrap_static_function() noexcept;
@@ -232,14 +232,14 @@ namespace Ubpa::UDRefl {
 
 	// Func: Ret(const? volatile? Object&, Args...)
 	// [result]
-	// - type : ({const?} void* obj, void* result_buffer, ArgPtrBuffer argptr_buffer) -> Destructor
+	// - type : void({const?} void* obj, void* result_buffer, ArgPtrBuffer argptr_buffer)
 	// - size : sizeof(Func)
 	template<typename Func>
 	constexpr auto wrap_member_function(Func&& func) noexcept;
 
 	// Func: Ret(Args...)
 	// [result]
-	// - type : (void* result_buffer, ArgPtrBuffer argptr_buffer) -> Destructor
+	// - type : void(void* result_buffer, ArgPtrBuffer argptr_buffer)
 	// - size : sizeof(Func)
 	template<typename Func>
 	constexpr auto wrap_static_function(Func&& func) noexcept;