+++ src/core/MethodPtr.cpp
@@ -2,34 +2,34 @@
 
 using namespace Ubpa::UDRefl;
 
-MethodPtr::MethodPtr(MemberVariableFunction* func, ResultDesc resultDesc, ParamList paramList) :
+MethodPtr::MethodPtr(MemberVariableFunction* func, Type result_type, ParamList paramList) :
 	func{ func },
-	resultDesc{ std::move(resultDesc) },
+	result_type{ std::move(result_type) },
 	paramList{ std::move(paramList) } { assert(func); }
 
-MethodPtr::MethodPtr(MemberConstFunction* func, ResultDesc resultDesc, ParamList paramList) :
+MethodPtr::MethodPtr(MemberConstFunction* func, Type result_type, ParamList paramList) :
 	func{ func },
-	resultDesc{ std::move(resultDesc) },
+	result_type{ std::move(result_type) },
 	paramList{ std::move(paramList) } { assert(func); }
 
-MethodPtr::MethodPtr(StaticFunction* func, ResultDesc resultDesc, ParamList paramList) :
+MethodPtr::MethodPtr(StaticFunction* func, Type result_type, ParamList paramList) :
 	func{ func },
-	resultDesc{ std::move(resultDesc) },
+	result_type{ std::move(result_type) },
 	paramList{ std::move(paramList) } { assert(func); }
 
-MethodPtr::MethodPtr(std::function<MemberVariableFunction> func, ResultDesc resultDesc, ParamList paramList) :
+MethodPtr::MethodPtr(std::function<MemberVariableFunction> func, Type result_type, ParamList paramList) :
 	func{ (assert(func), std::move(func)) },
-	resultDesc{ std::move(resultDesc) },
+	result_type{ std::move(result_type) },
 	paramList{ std::move(paramList) } {}
 
-MethodPtr::MethodPtr(std::function<MemberConstFunction> func, ResultDesc resultDesc, ParamList paramList) :
+MethodPtr::MethodPtr(std::function<MemberConstFunction> func, Type result_type, ParamList paramList) :
 	func{ (assert(func), std::move(func)) },
-	resultDesc{ std::move(resultDesc) },
+	result_type{ std::move(result_type) },
 	paramList{ std::move(paramList) } {}
 
-MethodPtr::MethodPtr(std::function<StaticFunction> func, ResultDesc resultDesc, ParamList paramList) :
+MethodPtr::MethodPtr(std::function<StaticFunction> func, Type result_type, ParamList paramList) :
 	func{ (assert(func), std::move(func)) },
-	resultDesc{ std::move(resultDesc) },
+	result_type{ std::move(result_type) },
 	paramList{ std::move(paramList) } {}
 
 MethodFlag MethodPtr::GetMethodFlag() const noexcept {
@@ -49,73 +49,61 @@ MethodFlag MethodPtr::GetMethodFlag() const noexcept {
 	}
 }
 
-Destructor MethodPtr::Invoke(void* obj, void* result_buffer, ArgPtrBuffer argptr_buffer) const {
-	return std::visit([=, this](const auto& f) {
+void MethodPtr::Invoke(void* obj, void* result_buffer, ArgPtrBuffer argptr_buffer) const {
+	std::visit([=, this](const auto& f) {
 		using Func = std::decay_t<decltype(f)>;
 		if constexpr (std::is_same_v<Func, MemberVariableFunction*>)
-			return f(obj, result_buffer, { argptr_buffer,paramList });
+			f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, MemberConstFunction*>)
-			return f(obj, result_buffer, { argptr_buffer,paramList });
+			f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, StaticFunction*>)
-			return f(     result_buffer, { argptr_buffer,paramList });
+			f(     result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, std::function<MemberVariableFunction>>)
-			return f(obj, result_buffer, { argptr_buffer,paramList });
+			f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, std::function<MemberConstFunction>>)
-			return f(obj, result_buffer, { argptr_buffer,paramList });
+			f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, std::function<StaticFunction>>)
-			return f(     result_buffer, { argptr_buffer,paramList });
+			f(     result_buffer, { argptr_buffer,paramList });
 		else
 			static_assert(always_false<Func>);
 	}, func);
 };
 
-Destructor MethodPtr::Invoke(const void* obj, void* result_buffer, ArgPtrBuffer argptr_buffer) const {
-	return std::visit([=, this](const auto& f)->Destructor {
+void MethodPtr::Invoke(const void* obj, void* result_buffer, ArgPtrBuffer argptr_buffer) const {
+	std::visit([=, this](const auto& f) {
 		using Func = std::decay_t<decltype(f)>;
-		if constexpr (std::is_same_v<Func, MemberVariableFunction*>) {
+		if constexpr (std::is_same_v<Func, MemberVariableFunction*>)
 			assert(false);
-			return {};
-		}
 		else if constexpr (std::is_same_v<Func, MemberConstFunction*>)
-			return f(obj, result_buffer, { argptr_buffer,paramList });
+			f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, StaticFunction*>)
-			return f(     result_buffer, { argptr_buffer,paramList });
-		else if constexpr (std::is_same_v<Func, std::function<MemberVariableFunction>>) {
+			f(     result_buffer, { argptr_buffer,paramList });
+		else if constexpr (std::is_same_v<Func, std::function<MemberVariableFunction>>)
 			assert(false);
-			return {};
-		}
 		else if constexpr (std::is_same_v<Func, std::function<MemberConstFunction>>)
-			return f(obj, result_buffer, { argptr_buffer,paramList });
+			f(obj, result_buffer, { argptr_buffer,paramList });
 		else if constexpr (std::is_same_v<Func, std::function<StaticFunction>>)
-			return f(     result_buffer, { argptr_buffer,paramList });
+			f(     result_buffer, { argptr_buffer,paramList });
 		else
 			static_assert(always_false<Func>);
 	}, func);
 };
 
-Destructor MethodPtr::Invoke(void* result_buffer, ArgPtrBuffer argptr_buffer) const {
-	return std::visit([=, this](const auto& f)->Destructor {
+void MethodPtr::Invoke(void* result_buffer, ArgPtrBuffer argptr_buffer) const {
+	std::visit([=, this](const auto& f) {
 		using Func = std::decay_t<decltype(f)>;
-		if constexpr (std::is_same_v<Func, MemberVariableFunction*>) {
+		if constexpr (std::is_same_v<Func, MemberVariableFunction*>)
 			assert(false);
-			return {};
-		}
-		else if constexpr (std::is_same_v<Func, MemberConstFunction*>) {
+		else if constexpr (std::is_same_v<Func, MemberConstFunction*>)
 			assert(false);
-			return {};
-		}
 		else if constexpr (std::is_same_v<Func, StaticFunction*>)
-			return f(result_buffer, { argptr_buffer,paramList });
-		else if constexpr (std::is_same_v<Func, std::function<MemberVariableFunction>>) {
+			f(result_buffer, { argptr_buffer,paramList });
+		else if constexpr (std::is_same_v<Func, std::function<MemberVariableFunction>>)
 			assert(false);
-			return {};
-		}
-		else if constexpr (std::is_same_v<Func, std::function<MemberConstFunction>>) {
+		else if constexpr (std::is_same_v<Func, std::function<MemberConstFunction>>)
 			assert(false);
-			return {};
-		}
 		else if constexpr (std::is_same_v<Func, std::function<StaticFunction>>)
-			return f(result_buffer, { argptr_buffer,paramList });
+			f(result_buffer, { argptr_buffer,paramList });
 		else
 			static_assert(always_false<Func>);
 	}, func);