+++ src/core/Object.cpp
@@ -17,7 +17,7 @@ ObjectView ObjectView::Var(Type base, Name field_name, FieldFlag flag) const {
 	return Mngr.Var(*this, base, field_name, flag);
 }
 
-InvokeResult ObjectView::Invoke(
+Type ObjectView::Invoke(
 	Name method_name,
 	void* result_buffer,
 	std::span<const Type> argTypes,
@@ -150,7 +150,7 @@ ObjectView ObjectView::AddRValueReference() const {
 	return { Mngr.tregistry.RegisterAddRValueReference(type), ptr };
 }
 
-InvocableResult ObjectView::IsInvocable(Name method_name, std::span<const Type> argTypes, MethodFlag flag) const {
+Type ObjectView::IsInvocable(Name method_name, std::span<const Type> argTypes, MethodFlag flag) const {
 	return Mngr.IsInvocable(type, method_name, argTypes, flag);
 }